+++ plugins/get_plugin_manifest.py
@@ -1,11 +1,27 @@
+print("hi")
 import urllib.parse
 import boto3
 import yaml
 import subprocess
 import sys
-from .utils import discover_manifest
+from npe2 import PluginManager
 
 s3 = boto3.client('s3')
+
+
+def discover_manifest(plugin_name):
+    pm = PluginManager()
+    pm.discover(include_npe1=False)
+    is_npe2 = True
+    try:
+        manifest = pm.get_manifest(plugin_name)
+    except KeyError:
+        pm.discover(include_npe1=True)
+        is_npe2 = False
+        # forcing lazy discovery to run
+        list(pm.iter_widgets())
+        manifest = pm.get_manifest(plugin_name)
+    return manifest, is_npe2
 
 
 def generate_manifest(event, context):
